Refactor to reduce size of BlockListener
