+++ src/test/java/world/bentobox/aoneblock/listeners/BlockListenerTest.java
@@ -1,15 +1,10 @@
 package world.bentobox.aoneblock.listeners;
 
-import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.never;
-import static org.mockito.Mockito.times;
-import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import java.beans.IntrospectionException;
@@ -18,9 +13,7 @@ import java.io.IOException;
 import java.lang.reflect.InvocationTargetException;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.util.ArrayList;
 import java.util.Comparator;
-import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 import java.util.concurrent.CompletableFuture;
@@ -61,7 +54,6 @@ import world.bentobox.bentobox.database.DatabaseSetup;
 import world.bentobox.bentobox.database.DatabaseSetup.DatabaseType;
 import world.bentobox.bentobox.database.objects.Island;
 import world.bentobox.bentobox.managers.IslandsManager;
-import world.bentobox.bentobox.managers.PlaceholdersManager;
 import world.bentobox.bentobox.managers.PlayersManager;
 import world.bentobox.bentobox.util.Util;
 import world.bentobox.level.Level;
@@ -103,8 +95,6 @@ public class BlockListenerTest {
     @Mock
     private Level level;
     @Mock
-    private PlaceholdersManager phm;
-    @Mock
     private Location location;
     @Mock
     private IslandsManager im;
@@ -140,9 +130,6 @@ public class BlockListenerTest {
         DatabaseType value = DatabaseType.JSON;
         when(plugin.getSettings()).thenReturn(pluginSettings);
         when(pluginSettings.getDatabaseType()).thenReturn(value);
-        when(plugin.getPlaceholdersManager()).thenReturn(phm);
-        // Placeholders
-        when(phm.replacePlaceholders(any(), anyString())).thenAnswer(a -> (String)a.getArgument(1, String.class));
 
         // Addon
         when(addon.getPlugin()).thenReturn(plugin);
@@ -206,34 +193,6 @@ public class BlockListenerTest {
     }
 
     /**
-     * Test method for {@link world.bentobox.aoneblock.listeners.BlockListener#replacePlaceholders(org.bukkit.entity.Player, java.lang.String, java.lang.String, world.bentobox.bentobox.database.objects.Island, java.util.List)}.
-     */
-    @Test
-    public void testReplacePlaceholders() {
-        // Commands
-        /*
-         * [island] - Island name
-         * [owner] - Island owner's name
-         * [player] - The name of the player who broke the block triggering the commands
-         * [phase] - the name of this phase
-         * [blocks] - the number of blocks broken
-         * [level] - island level (Requires Levels Addon)
-         * [bank-balance] - island bank balance (Requires Bank Addon)
-         * [eco-balance] - player's economy balance (Requires Vault and an economy plugin)
-
-         */
-        List<String> commandList = new ArrayList<>();
-
-        commandList.add("no replacement");
-        commandList.add("[island] [owner] [phase] [blocks] [level] [bank-balance] [eco-balance]");
-        List<String> r = bl.replacePlaceholders(player, "phaseName", "1000", island, commandList);
-        assertEquals(2, r.size());
-        assertEquals("no replacement", r.get(0));
-        assertEquals("island_name tastybento2 phaseName 1000 1000 100000.0 0.0", r.get(1));
-        verify(phm, times(2)).replacePlaceholders(eq(player), any());
-    }
-
-    /**
      * Test method for {@link world.bentobox.aoneblock.listeners.BlockListener#onBlockFromTo(org.bukkit.event.block.BlockFromToEvent)}
      */
     @Test
@@ -281,143 +240,5 @@ public class BlockListenerTest {
         bl.onBlockFromTo(e);
         assertTrue(e.isCancelled());
     }
-    
-    /**
-     * Test method for {@link world.bentobox.aoneblock.listeners.BlockListener#checkPhase(Player, Island, world.bentobox.aoneblock.dataobjects.OneBlockIslands, world.bentobox.aoneblock.oneblocks.OneBlockPhase)}
-     */
-    @Test
-    public void testCheckPhase() {
-        // Set up that a phase has been completed
-        is = new OneBlockIslands(UUID.randomUUID().toString());
-        is.setPhaseName("Previous");
-        is.setBlockNumber(500);
-        is.setLifetime(500L);
-        // The phase the user has just moved to
-        phase = new OneBlockPhase("500");
-        phase.setPhaseName("Next Phase");
-        phase.setStartCommands(List.of("start1", "start2"));
-        
-        // The previous phase        
-        OneBlockPhase previous = mock(OneBlockPhase.class);
-        when(previous.getPhaseName()).thenReturn("Previous");
-        
-        when(obm.getPhase("Previous")).thenReturn(Optional.of(previous));
-        
-        assertTrue(bl.checkPhase(player, island, is, phase));
-        // Verify commands run
-        verify(previous).getEndCommands();
-        verify(previous).getFirstTimeEndCommands();
-        // Verify title shown
-        verify(player).sendTitle("Next Phase", null, -1, -1, -1);
-        
-    }
-    
-    /**
-     * Test method for {@link world.bentobox.aoneblock.listeners.BlockListener#checkPhase(Player, Island, world.bentobox.aoneblock.dataobjects.OneBlockIslands, world.bentobox.aoneblock.oneblocks.OneBlockPhase)}
-     */
-    @Test
-    public void testCheckPhaseSecondTime() {
-        // Set up that a phase has been completed
-        is = new OneBlockIslands(UUID.randomUUID().toString());
-        is.setPhaseName("Previous");
-        is.setBlockNumber(500);
-        is.setLifetime(10500L);
-        // The phase the user has just moved to
-        phase = new OneBlockPhase("500");
-        phase.setPhaseName("Next Phase");
-        phase.setStartCommands(List.of("start1", "start2"));
-        
-        // The previous phase        
-        OneBlockPhase previous = mock(OneBlockPhase.class);
-        when(previous.getPhaseName()).thenReturn("Previous");
-        
-        when(obm.getPhase("Previous")).thenReturn(Optional.of(previous));
-        
-        assertTrue(bl.checkPhase(player, island, is, phase));
-        // Verify commands run
-        verify(previous).getEndCommands();
-        verify(previous, never()).getFirstTimeEndCommands();
-        // Verify title shown
-        verify(player).sendTitle("Next Phase", null, -1, -1, -1);
-        
-    }
-    
-    /**
-     * Test method for {@link world.bentobox.aoneblock.listeners.BlockListener#checkPhase(Player, Island, world.bentobox.aoneblock.dataobjects.OneBlockIslands, world.bentobox.aoneblock.oneblocks.OneBlockPhase)}
-     */
-    @Test
-    public void testCheckPhaseNullPlayer() {
-        // Set up that a phase has been completed
-        is = new OneBlockIslands(UUID.randomUUID().toString());
-        is.setPhaseName("Previous");
-        is.setBlockNumber(500);
-        is.setLifetime(500L);
-        // The phase the user has just moved to
-        phase = new OneBlockPhase("500");
-        phase.setPhaseName("Next Phase");
-        phase.setStartCommands(List.of("start1", "start2"));
-        
-        // The previous phase        
-        OneBlockPhase previous = mock(OneBlockPhase.class);
-        when(previous.getPhaseName()).thenReturn("Previous");
-        
-        when(obm.getPhase("Previous")).thenReturn(Optional.of(previous));
-        
-        assertTrue(bl.checkPhase(null, island, is, phase));
-        // Verify commands run
-        verify(previous).getEndCommands();
-        verify(previous).getFirstTimeEndCommands();
-        // Verify title shown
-        verify(player).sendTitle("Next Phase", null, -1, -1, -1);
-        
-    }
-    
-    /**
-     * Test method for {@link world.bentobox.aoneblock.listeners.BlockListener#checkPhase(Player, Island, world.bentobox.aoneblock.dataobjects.OneBlockIslands, world.bentobox.aoneblock.oneblocks.OneBlockPhase)}
-     */
-    @Test
-    public void testCheckPhaseNPCPlayer() {
-        when(player.hasMetadata("NPC")).thenReturn(true);
-        // Set up that a phase has been completed
-        is = new OneBlockIslands(UUID.randomUUID().toString());
-        is.setPhaseName("Previous");
-        is.setBlockNumber(500);
-        is.setLifetime(500L);
-        // The phase the user has just moved to
-        phase = new OneBlockPhase("500");
-        phase.setPhaseName("Next Phase");
-        phase.setStartCommands(List.of("start1", "start2"));
-        
-        // The previous phase        
-        OneBlockPhase previous = mock(OneBlockPhase.class);
-        when(previous.getPhaseName()).thenReturn("Previous");
-        
-        when(obm.getPhase("Previous")).thenReturn(Optional.of(previous));
-        
-        assertTrue(bl.checkPhase(player, island, is, phase));
-        // Verify commands run
-        verify(previous).getEndCommands();
-        verify(previous).getFirstTimeEndCommands();
-        // Verify title shown
-        verify(player).sendTitle("Next Phase", null, -1, -1, -1);
-        
-    }
-    
-    /**
-     * Test method for {@link world.bentobox.aoneblock.listeners.BlockListener#checkPhase(Player, Island, world.bentobox.aoneblock.dataobjects.OneBlockIslands, world.bentobox.aoneblock.oneblocks.OneBlockPhase)}
-     */
-    @Test
-    public void testCheckSamePhase() {
-        is = new OneBlockIslands(UUID.randomUUID().toString());
-        is.setPhaseName("Previous");
-        is.setBlockNumber(500);
-        is.setLifetime(500L);
-        // The phase the user has just moved to
-        phase = new OneBlockPhase("500");
-        phase.setPhaseName("Previous");
-                
-        assertFalse(bl.checkPhase(player, island, is, phase));
-        
-    }
-    
+
 }
